Move code that depends on Machine and Executive to own module
