+++ ethcore/src/lib.rs
@@ -171,6 +171,7 @@ pub mod verification;
 
 mod externalities;
 mod tx_filter;
+mod executive_state;
 
 #[cfg(test)]
 mod tests;
